Metade do Clean Code
Pendente apenas comentar o código, deixando tudo bem explicado.
